+++ packages/api/tests/lib/test-granules.js
@@ -2,10 +2,14 @@ const test = require('ava');
 const sinon = require('sinon');
 const proxyquire = require('proxyquire');
 const { randomId } = require('@cumulus/common/test-utils');
+const { bootstrapElasticSearch } = require('@cumulus/es-client/bootstrap');
+const { Search } = require('@cumulus/es-client/search');
+const indexer = require('@cumulus/es-client/indexer');
 
 const {
   getExecutionProcessingTimeInfo,
   moveGranuleFilesAndUpdateDatastore,
+  granuleEsQuery,
 } = require('../../lib/granules');
 const { fakeGranuleFactoryV2 } = require('../../lib/testUtils');
 
@@ -263,7 +267,7 @@ test.serial('getGranulesForPayload returns unique granules from query', async (t
   );
 });
 
-test.serial('getGranulsForPayload handles query paging', async (t) => {
+test.serial('getGranulesForPayload handles query paging', async (t) => {
   const granuleId1 = randomId('granule');
   const granuleId2 = randomId('granule');
   const granuleId3 = randomId('granule');
@@ -343,4 +347,63 @@ test('translateGranule() will translate an old-style granule file and numeric pr
     }
   );
   t.is(translatedGranule.productVolume, oldProductVolume.toString());
+});
+
+test.serial('granuleEsQuery returns if the query has a bad timestamp and responseQueue.body.hits.total.value is 0', async (t) => {
+  const esAlias = randomId('esAlias');
+  const esIndex = randomId('esindex');
+  process.env.ES_INDEX = esAlias;
+  const esClient = await Search.es();
+
+  await bootstrapElasticSearch({
+    host: 'fakeHost',
+    index: esIndex,
+    alias: esAlias,
+  });
+  const granuleId = randomId();
+
+  const fakeGranule = fakeGranuleFactoryV2({ granuleId });
+  await indexer.indexGranule(esClient, fakeGranule, esIndex);
+  const esGranulesClient = new Search(
+    {},
+    'granule',
+    process.env.ES_INDEX
+  );
+
+  const query = {
+    query: {
+      bool: {
+        must: [],
+        filter: [
+          {
+            range: {
+              '@timestamp': {
+                gte: '2022-11-07T23:59:00.220Z',
+                lte: '2022-11-08T19:51:36.220Z',
+                format: 'strict_date_optional_time',
+              },
+            },
+          },
+        ],
+        should: [],
+        must_not: [],
+      },
+    },
+  };
+  t.like(await esGranulesClient.get(granuleId), fakeGranule);
+  const testBodyHits = {
+    total: {
+      value: 0,
+      relation: 'eq',
+    },
+    max_score: null,
+    hits: [],
+  };
+
+  t.truthy(await granuleEsQuery({
+    index: esIndex,
+    query,
+    source: ['granuleId'],
+    testBodyHits,
+  }));
 });